# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ZEINA Hyperviseur API
  version: "1.0.0"
  description: |
    REST + WebSocket API du superviseur IoT ZEINA.

    ## Authentification
    1. POST `/v1/auth/login` retourne un access token (15 min) + un refresh token (cookie httpOnly, 7 j).
    2. Les routes protégées attendent l'access token dans `Authorization: Bearer <token>`.
    3. Quand le token expire, POST `/v1/auth/refresh` produit un nouvel access (utilise le cookie).

    ## Modèle d'autorisation
    - **Superadmin** (`is_superadmin=true`) : bypass complet.
    - **Owner du tenant** (`tenant_role=owner`) : accès implicite à tous les sites + gestion users/rôles.
    - **Membre simple** : accès aux sites où il est listé dans `site_members`, selon le rôle.
    - Chaque rôle attribue un niveau (`none`/`read`/`write`) sur 4 fonctionnalités : `dashboard`, `devices`, `rules`, `members`.

    ## Test rapide
    1. Clique « Authorize » en haut, paste l'access token (sans le préfixe Bearer).
    2. Tous les calls authentifiés vont passer.
servers:
  - url: /
    description: |
      Origine courante (relatif). Quand tu importes ce spec dans Bruno/Postman,
      remplace par `http://localhost:5173` (via nginx) ou `http://localhost:3000`
      (API directe) — sans slash final.
...
security:
  - bearerAuth: []
paths:
...
  /v1/sites/{siteId}/alarms:
    get:
      tags: [Alarms]
      summary: Liste les alarmes d'un site
      description: |
        Retourne les alarmes filtrées par état. États possibles :
          - `active` (défaut) : `triggered` ∪ `acknowledged`
          - `triggered` / `acknowledged` / `resolved` / `archived`
          - `all` : toutes
      parameters:
        - { $ref: "#/components/parameters/SiteId" }
        - in: query
          name: state
          schema: { type: string, enum: [active, triggered, acknowledged, resolved, archived, all] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 200, minimum: 1, maximum: 1000 }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Alarm" } } } } }
  /v1/sites/{siteId}/alarms/counts:
    get:
      tags: [Alarms]
      summary: Compteurs d'alarmes par état (onglets UI)
      parameters:
        - { $ref: "#/components/parameters/SiteId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  triggered:    { type: integer }
                  acknowledged: { type: integer }
                  resolved:     { type: integer }
                  archived:     { type: integer }
                  active:       { type: integer, description: "triggered + acknowledged" }
                  all:          { type: integer }
  /v1/alarms/{id}:
    get:
      tags: [Alarms]
      summary: Détail d'une alarme
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Alarm" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/alarms/{id}/events:
    get:
      tags: [Alarms]
      summary: Historique des évènements d'une alarme
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AlarmEvent" } } } } }
  /v1/alarms/{id}/comments:
    get:
      tags: [Alarms]
      summary: Liste les commentaires d'une alarme
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AlarmComment" } } } } }
    post:
      tags: [Alarms]
      summary: Ajoute un commentaire à une alarme
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AlarmComment" } } } }
  /v1/alarms/{id}/acknowledge:
    post:
      tags: [Alarms]
      summary: Prendre en compte une alarme (passe à `acknowledged`)
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK }
  /v1/alarms/{id}/resolve:
    post:
      tags: [Alarms]
      summary: Acquitter une alarme (passe à `resolved`)
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK }
  /v1/alarms/{id}/archive:
    post:
      tags: [Alarms]
      summary: Archiver une alarme
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK }

  # ---------------------------------------------------------------------------
  # AUDIT  (owner / superadmin)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: ID du site
...
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string, example: validation }
        message: { type: string, example: "email and password are required" }
        details: { type: object, additionalProperties: true }
...
    Alarm:
      type: object
      properties:
        id:                 { type: string, format: uuid }
        tenant_id:          { type: string, format: uuid }
        site_id:            { type: string, format: uuid }
        rule_id:            { type: string, format: uuid }
        rule_name:          { type: string }
        device_id:          { type: string, format: uuid, nullable: true }
        device_slug:        { type: string, nullable: true }
        device_name:        { type: string, nullable: true }
        zone_id:            { type: string, format: uuid, nullable: true }
        zone_name:          { type: string, nullable: true }
        label:              { type: string, example: "Dépassement de seuil" }
        name:               { type: string }
        description:        { type: string, nullable: true }
        severity:           { type: string, enum: [minor, major, critical] }
        model:              { type: string, default: "Standard" }
        status_text:        { type: string, nullable: true, example: "Comportement anormal" }
        state:              { type: string, enum: [triggered, acknowledged, resolved, archived] }
        attribute:          { type: string, nullable: true, example: "temperature" }
        trigger_count:      { type: integer }
        last_value:         { type: number, nullable: true }
        unit:               { type: string, nullable: true }
        opened_at:          { type: string, format: date-time }
        last_triggered_at:  { type: string, format: date-time }
        acked_at:           { type: string, format: date-time, nullable: true }
        resolved_at:        { type: string, format: date-time, nullable: true }
        archived_at:        { type: string, format: date-time, nullable: true }
        ack_user_email:     { type: string, nullable: true }
        resolve_user_email: { type: string, nullable: true }
    AlarmEvent:
      type: object
      properties:
        id:            { type: string, format: uuid }
        alarm_id:      { type: string, format: uuid }
        ts:            { type: string, format: date-time }
        state:         { type: string, enum: [triggered, acknowledged, resolved, archived] }
        severity:      { type: string, enum: [minor, major, critical] }
        description:   { type: string, nullable: true }
        trigger_count: { type: integer, nullable: true }
        value:         { type: number, nullable: true }
        user_email:    { type: string, nullable: true }
    AlarmComment:
      type: object
      properties:
        id:         { type: string, format: uuid }
        alarm_id:   { type: string, format: uuid }
        user_email: { type: string, nullable: true }
        body:       { type: string }
        created_at: { type: string, format: date-time }